# improve validator speed by caching tag parameters