improve validator speed by caching tag parameters

This factors out the type checking and the parameter parsing (including
regexp compilation) from the hot path. Checking UUIDs with a regexp is
still pretty slow though - we could use a custom validator (e.g. fastuuid.ValidUUID)
to make things faster still, but it's probably not a good idea to take on
another external dependency until modules are enabled.

From 798f45bdc3c0724e2ca3526ae7da57f80b392a6d (with type caching) to this:

        name        old time/op  new time/op  delta
        Latitude-4   309ns ± 1%   274ns ± 1%  -11.17%  (p=0.000 n=5+4)
        UUID-4      22.0µs ±11%   1.0µs ±15%  -95.49%  (p=0.008 n=5+5)
        Validate-4  1.00µs ± 3%  0.63µs ± 4%  -37.04%  (p=0.008 n=5+5)

From e309c54f7b0b72f0b478f50cc70778b198d2f79f (without type caching) to this:

        name        old time/op  new time/op  delta
        Latitude-4  3.58µs ±13%  0.27µs ± 1%  -92.32%  (p=0.016 n=5+4)
        UUID-4      23.8µs ± 5%   1.0µs ±15%  -95.85%  (p=0.008 n=5+5)
        Validate-4  13.3µs ±10%   0.6µs ± 4%  -95.27%  (p=0.008 n=5+5)